"refactor: 1. API 명세서에 따른 URL 매핑 수정, 2. API 명세에 따른 admin 컨트롤러 분리"
