+++ http/request.http
@@ -1,5 +1,5 @@
-### 쿠폰 정책 생성
-POST http://localhost:8080/coupon-policies
+### 쿠폰 정책 생성 -> * 관리자용 *
+POST http://localhost:8080/admin/coupon-policies
 Content-Type: application/json
 
 {
@@ -9,21 +9,21 @@ Content-Type: application/json
   "discountAmount": 3000
 }
 
-### 쿠폰 정책 단건 조회 {id}
-GET http://localhost:8080/coupon-policies/1
+### 쿠폰 정책 단건 조회 {id} -> * 관리자용 *
+GET http://localhost:8080/admin/coupon-policies/1
 
-### 쿠폰 정책 단건 조회 {code}
-GET http://localhost:8080/coupon-policies?code=AMT15K_DC3K
+### 쿠폰 정책 단건 조회 {code} -> * 관리자용 *
+GET http://localhost:8080/admin/coupon-policies?code=AMT15K_DC3K
 
-### 쿠폰 정책 전체 조회
-GET http://localhost:8080/coupon-policies
+### 쿠폰 정책 전체 조회 -> * 관리자용 *
+GET http://localhost:8080/admin/coupon-policies
 
 
-### 쿠폰 정책 삭제 (= 비활성화)
-DELETE http://localhost:8080/coupon-policies/AMT15K_DC3K
+### 쿠폰 정책 삭제 (= 비활성화) -> * 관리자용 *
+DELETE http://localhost:8080/admin/coupon-policies/AMT15K_DC3K
 
-### 쿠폰 정책 수정
-PUT http://localhost:8080/coupon-policies/AMT15K_DC3K
+### 쿠폰 정책 수정 -> * 관리자용 *
+PUT http://localhost:8080/admin/coupon-policies/AMT15K_DC3K
 Content-Type: application/json
 
 {
@@ -33,47 +33,47 @@ Content-Type: application/json
 
 
 
-### (정책기반) -> 쿠폰 생성
-POST http://localhost:8080/coupons
+### (정책기반=) 쿠폰 생성 -> * 관리자용 *
+POST http://localhost:8080/admin/coupons
 Content-Type: application/json
 
 {
-  "couponName": "도서 할인 쿠폰",
+  "couponName": "일반 쿠폰",
   "policyCode": "AMT15K_DC3K",
+  "couponType": "GENERAL",
   "validFrom": "2025-01-01",
   "validTo": "2026-01-01",
-  "couponType": "BOOKS",
-  "comment": "특정 도서 할인쿠폰 단 10장",
-  "issueCount": 10,
-  "bookName": "자바"
+  "comment": "최소주문 금액 15,000 이상 시 3000원 할인",
+  "issueCount": 10
 }
 
-### (정책기반) -> 쿠폰 단건 조회 {id} // id ->
-GET http://localhost:8080/coupons/14
+### (정책기반=) 쿠폰 단건 조회 {id} -> * 관리자용 *
+GET http://localhost:8080/admin/coupons/1
 
-### (정책기반) -> 쿠폰 (policyCode)목록 조회 {policyCode}
-GET http://localhost:8080/coupons?policyCode=AMT15K_DC333K
+### (정책기반=) 쿠폰 단건 조회 {policyCode} -> 관리자쪽에서는 조회에 대한 url이 많으면 안될것같아서 추후 삭제 하거나 수정예정
+GET http://localhost:8080/coupons?policyCode=AMT15K_DC3K
 
-### (정책기반) -> 쿠폰 (type)목록 조회 {type} || WELCOME, BIRTHDAY, GENERAL
-GET http://localhost:8080/coupons?type=BIRTHDAY
+### (정책기반=) 쿠폰 (type)목록 조회 {type} || WELCOME, BIRTHDAY, GENERAL, BOOKS, CATEGORY
+GET http://localhost:8080/coupons?type=GENERAL
 
-### (정책기반) -> 쿠폰 (name)목록 조회 {name}
-GET http://localhost:8080/coupons?name=웰컴쿠폰
+### (정책기반=) 쿠폰 (name)목록 조회 {name}
+GET http://localhost:8080/coupons?name=일반 쿠폰
 
-### (정책기반) -> 쿠폰 전체조회
+### (정책기반=) 쿠폰 전체조회
 GET http://localhost:8080/coupons
 
-### (정책기반) -> 쿠폰 수정
-PUT http://localhost:8080/coupons/1
+### (정책기반=) 쿠폰 수정 -> * 관리자용 *
+PUT http://localhost:8080/admin/coupons/1
 Content-Type: application/json
 
 {
-  "couponName": "수정된 신규가입쿠폰",
-  "comment": "50,000 이상 구매 시 50% 할인"
+  "couponName": "수정된 일반 쿠폰",
+  "comment": "쿠폰 수량 +30 추가!",
+  "issueCount": 40
 }
 
 ### (정책기반) -> 쿠폰 삭제
-DELETE http://localhost:8080/coupons/1
+DELETE http://localhost:8080/admin/coupons/1
 
 ### (임시) -> 회원 생성
 POST http://localhost:8080/members
@@ -87,17 +87,17 @@ Content-Type: application/json
 
 
 ### (회원) -> 쿠폰 발급
-POST http://localhost:8080/member-coupons
+POST http://localhost:8080/members/member-coupons
 Content-Type: application/json
 
 {
   "memberEmail": "test@test.com",
-  "couponName": "도서 할인 쿠폰"
+  "couponName": "수정된 일반 쿠폰"
 }
 
 ### (회원) -> 쿠폰 발급내역
-GET http://localhost:8080/member-coupons/test@test.com
+GET http://localhost:8080/members/member-coupons/test@test.com
 
 
 ### (회원) -> 쿠폰 사용
-PUT http://localhost:8080/member-coupons/test@test.com/2/use
+PUT http://localhost:8080/members/member-coupons/test@test.com/1/use
