+++ src/main/java/com/nhnacademy/illuwa/domain/coupons/controller/CouponController.java
@@ -19,33 +19,6 @@ public class CouponController {
 
     private final CouponService couponService;
 
-    // (정책기반) 쿠폰 생성
-    @PostMapping
-    public ResponseEntity<CouponCreateResponse> createCoupon(@RequestBody @Valid CouponCreateRequest request) {
-//        if (Objects.isNull(request.getBookName())) {
-//            CouponCreateResponse response = couponService.createCoupon(request);
-//            return new ResponseEntity<>(response, HttpStatus.CREATED);
-//        } else {
-//            CouponCreateResponse response = couponService.createCouponByBookTitle(request.getBookName(), request);
-//            return new ResponseEntity<>(response, HttpStatus.CREATED);
-//        }
-        CouponCreateResponse response = couponService.createCoupon(request);
-        return new ResponseEntity<>(response, HttpStatus.CREATED);
-    }
-//
-//    @PostMapping
-//    public ResponseEntity<CouponCreateResponse> createCouponWithBook(@RequestBody @Valid CouponCreateRequest request) {
-//        CouponCreateResponse response = couponService.createCouponByBookTitle(request.getBookName(), request);
-//        return new ResponseEntity<>(response, HttpStatus.CREATED);
-//    }
-
-    // (정책기반) 쿠폰 단건 조회 {id}
-    @GetMapping("/{id}")
-    public ResponseEntity<CouponResponse> getCoupon(@PathVariable Long id) {
-        CouponResponse response = couponService.getCouponById(id);
-        return ResponseEntity.ok(response);
-    }
-
     // (정책기반) 쿠폰 목록( =PolicyCode로 조회)
     // Query Parameter 방식
     @GetMapping(params = "policyCode")
@@ -76,21 +49,5 @@ public class CouponController {
     public ResponseEntity<List<CouponResponse>> getAllCoupons() {
         List<CouponResponse> response = couponService.getAllCoupons();
         return ResponseEntity.ok(response);
-    }
-
-    // (정책기반) 쿠폰 수정
-    @PutMapping("{id}")
-    public ResponseEntity<CouponUpdateResponse> updateCoupon(
-            @PathVariable Long id,
-            @RequestBody CouponUpdateRequest request) {
-        CouponUpdateResponse response = couponService.updateCoupon(id, request);
-        return ResponseEntity.ok(response);
-    }
-
-    // (정책기반) 쿠폰 삭제
-    @DeleteMapping("{id}")
-    public ResponseEntity<Void> deleteCoupon(@PathVariable Long id) {
-        couponService.deleteCoupon(id);
-        return ResponseEntity.noContent().build();
     }
 }
