+++ src/main/java/com/nhnacademy/illuwa/domain/coupons/controller/CouponPolicyController.java
@@ -12,7 +12,7 @@ import java.util.List;
 
 @RestController
 @AllArgsConstructor
-@RequestMapping("/coupon-policies")
+@RequestMapping("admin/coupon-policies")
 public class CouponPolicyController {
 
     private final CouponPolicyService couponPolicyService;
