+++ src/main/java/com/nhnacademy/illuwa/domain/coupons/external/book/BookDto.java
@@ -9,5 +9,6 @@ public class BookDto {
     Long bookId;
     String title;
     String category;
-    String isbn;
+    // 도서 존재 O -> 쿠폰을 생성  (도서1) - book -> 1
+    // 주문 -> 주문이 도서1 도서2 도서3 -> 주문 처리를하는데 쿠폰이 활성화 (도서1 -> 도서2) id -2  useCoupon()
 }
